+++ l10n_ro_stock_account/models/account_invoice.py
@@ -45,48 +45,6 @@ class AccountMove(models.Model):
 
         return res
 
-    def is_reception_notice(self):
-        self.ensure_one()
-        purchases = self.line_ids.mapped("purchase_line_id.order_id")
-        picking_notice = self.env["stock.picking"].search(
-            [
-                ("id", "in", purchases.mapped("picking_ids").ids),
-                ("state", "=", "done"),
-                ("notice", "=", True),
-            ]
-        )
-        if picking_notice:
-            return True
-        return False
-
-    def get_reception_account(self):
-        self.ensure_one()
-        account = self.env["account.account"]
-        acc_payable = self.company_id.property_stock_picking_payable_account_id
-        valuation_stock_moves = self.env["stock.move"].search(
-            [
-                (
-                    "purchase_line_id",
-                    "in",
-                    self.line_ids.mapped("purchase_line_id").ids,
-                ),
-                ("state", "=", "done"),
-                ("picking_id.notice", "=", True),
-                ("product_qty", "!=", 0.0),
-            ]
-        )
-        if valuation_stock_moves:
-            acc_moves = valuation_stock_moves.mapped("account_move_ids")
-            lines = self.env["account.move.line"].search(
-                [("move_id", "in", acc_moves.ids)]
-            )
-            lines_diff_acc = lines.mapped("account_id").filtered(
-                lambda a: a != acc_payable
-            )
-            if lines_diff_acc:
-                account = lines_diff_acc[0]
-        return account
-
 
 class AccountMoveLine(models.Model):
     _inherit = "account.move.line"
@@ -116,27 +74,6 @@ class AccountMoveLine(models.Model):
         return valuation_stock_moves
 
     def _get_computed_account(self):
-        if (
-            self.product_id.type == "product"
-            and self.move_id.company_id.anglo_saxon_accounting
-        ):
-            if self.move_id.is_purchase_document():
-                purchase = self.purchase_order_id
-                if purchase and self.product_id.purchase_method == "receive":
-                    # Control bills based on received quantities
-                    if any(
-                        [p.notice or p._is_dropshipped() for p in purchase.picking_ids]
-                    ):
-                        self = self.with_context(valued_type="invoice_in_notice")
-            if self.move_id.is_sale_document():
-                sales = self.sale_line_ids
-                if sales and self.product_id.invoice_policy == "delivery":
-                    # Control bills based on received quantities
-                    sale = self.sale_line_ids[0].order_id
-                    if any(
-                        [p.notice and not p._is_dropshipped() for p in sale.picking_ids]
-                    ):
-                        self = self.with_context(valued_type="invoice_out_notice")
         res = super(AccountMoveLine, self)._get_computed_account()
         # Take accounts from stock location in case the category allow changinc
         # accounts and the picking is not notice
@@ -147,18 +84,14 @@ class AccountMoveLine(models.Model):
         ):
             fiscal_position = self.move_id.fiscal_position_id
             if self.move_id.is_purchase_document():
-                stock_moves = self.purchase_line_id.move_ids.filtered(
-                    lambda sm: not sm.picking_id.notice
-                )
+                stock_moves = self.purchase_line_id.move_ids
                 for stock_move in stock_moves.filtered(lambda m: m.state == "done"):
                     if stock_move.location_dest_id.property_stock_valuation_account_id:
                         location = stock_move.location_dest_id
                         res = location.property_stock_valuation_account_id
             if self.move_id.is_sale_document():
                 sales = self.sale_line_ids.filtered(lambda s: s.move_ids)
-                for stock_move in sales.move_ids.filtered(
-                    lambda m: not m.picking_id.notice and m.state == "done"
-                ):
+                for stock_move in sales.move_ids:
                     if stock_move.location_id.property_account_income_location_id:
                         location = stock_move.location_id
                         res = location.property_account_income_location_id
