+++ l10n_ro_stock_account/tests/test_purchase.py
@@ -88,21 +88,3 @@ class TestStockPurchase(TestStockCommon):
 
         # verificare inregistrare diferenta de pret
         self.check_account_diff(0, 0)
-
-    def test_nir_with_notice_and_invoice(self):
-        """
-        Receptie produse pe baza de aviz si inregistare
-        ulterioara a facturii
-        """
-        self.create_po(notice=True)
-
-        self.check_stock_valuation(self.val_p1_i, self.val_p2_i)
-        self.check_account_valuation(self.val_p1_i, self.val_p2_i)
-
-        self.create_invoice()
-
-        self.check_stock_valuation(self.val_p1_i, self.val_p2_i)
-        self.check_account_valuation(self.val_p1_i, self.val_p2_i)
-
-        # soldul lui 408 trebuie sa fie zero
-        self.check_account_valuation(0, 0, self.stock_picking_payable_account_id)
