+++ l10n_ro_stock_account_notice/__init__.py
@@ -1,0 +1,1 @@
+from . import models
