+++ l10n_ro_stock_account_notice/__manifest__.py
@@ -1,17 +1,16 @@
 # Copyright (C) 2022 NextERP Romania
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 {
-    "name": "Romania - Stock Accounting Price Difference",
+    "name": "Romania - Stock Accounting Notice",
     "version": "14.0.1.1.0",
     "category": "Localization",
-    "summary": "Romania - Stock Accounting Price Difference",
+    "summary": "Romania - Stock Accounting Notice",
     "author": "NextERP Romania," "Dorin Hongu," "Odoo Community Association (OCA)",
     "website": "https://github.com/OCA/l10n-romania",
-    "depends": ["l10n_ro_stock_account_notice", "stock_landed_costs"],
+    "depends": ["l10n_ro_stock_account"],
     "license": "AGPL-3",
     "data": [
-        "wizard/price_difference_confirmation.xml",
-        "security/ir.model.access.csv",
+        "views/stock_picking_view.xml",
     ],
     "installable": True,
     "auto_install": True,
