+++ l10n_ro_stock_account_notice/models/__init__.py
@@ -1,0 +1,5 @@
+from . import account_move
+from . import account_move_line
+from . import product
+from . import stock_move
+from . import stock_picking
