+++ l10n_ro_stock_account_notice/models/stock_picking.py
@@ -1,0 +1,21 @@
+# Copyright (C) 2014 Forest and Biomass Romania
+# Copyright (C) 2020 NextERP Romania
+# Copyright (C) 2020 Terrabit
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
+
+from odoo import fields, models
+
+
+class StockPicking(models.Model):
+    _inherit = "stock.picking"
+
+    # Prin acest camp se indica daca un produs care e stocabil trece prin
+    # contul 408 / 418 la achizitie sau vanzare
+    # receptie/ livrare in baza de aviz
+    notice = fields.Boolean(
+        "Is a notice",
+        states={"done": [("readonly", True)], "cancel": [("readonly", True)]},
+        default=False,
+        help="With this field the reception/delivery is set as a notice. "
+        "The generated account move will contain accounts 408/418.",
+    )
