+++ l10n_ro_stock_account_notice/readme/CONTRIBUTORS.rst
@@ -1,0 +1,12 @@
+* `NextERP Romania <https://www.nexterp.ro>`_:
+
+  * Fekete Mihai <feketemihai@nexterp.ro>
+  * Cojocaru Marcel <cojocarumarcel@nexterp.ro>
+
+* `Terrabit <https://www.terrabit.ro>`_:
+
+  * Dorin Hongu <dhongu@gmail.com>
+
+
+
+Do not contact contributors directly about support or help with technical issues.
