+++ l10n_ro_stock_account_notice/readme/DESCRIPTION.rst
@@ -1,0 +1,1 @@
+The module is managing price diferences between invoices and receptions. It will create a landed cost with a "Price Difference" product available in the configuration.
