+++ l10n_ro_stock_account_notice/tests/__init__.py
@@ -1,0 +1,3 @@
+from . import test_retur
+from . import test_purchase
+from . import test_sale
