+++ l10n_ro_stock_account_notice/tests/test_purchase.py
@@ -1,0 +1,29 @@
+# Copyright (C) 2020 Terrabit
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
+
+from odoo.tests import tagged
+
+from odoo.addons.l10n_ro_stock_account.tests.common import TestStockCommon
+
+# Generare note contabile la achizitie
+
+
+@tagged("post_install", "-at_install")
+class TestStockPurchase(TestStockCommon):
+    def test_nir_with_notice_and_invoice(self):
+        """
+        Receptie produse pe baza de aviz si inregistare
+        ulterioara a facturii
+        """
+        self.create_po(vals={"notice": True})
+
+        self.check_stock_valuation(self.val_p1_i, self.val_p2_i)
+        self.check_account_valuation(self.val_p1_i, self.val_p2_i)
+
+        self.create_invoice()
+
+        self.check_stock_valuation(self.val_p1_i, self.val_p2_i)
+        self.check_account_valuation(self.val_p1_i, self.val_p2_i)
+
+        # soldul lui 408 trebuie sa fie zero
+        self.check_account_valuation(0, 0, self.stock_picking_payable_account_id)
